+++ init/reiniciar_router.sh
@@ -18,47 +18,53 @@
 # Usage         : /etc/init.d/reiniciar_router {start|stop|restart|status}
 # Notes         :
 
+if [ "$UID" -ne '0' ]; then
+	echo 'Este script debe ser lanzado con permisos de root. ¿Quizá anteponiéndole la orden sudo?'
 
-nombre=reiniciar_router
-directorio='/opt/RPPGCT'
+	exit -1
 
-case "$1" in
-
-  start)
-    if [ -f /var/lock/${nombre}.lock ]; then
-      echo "${nombre}.py ya está en ejecucuón"
-    else
-      echo "Iniciando ${nombre}.py"
-      ${directorio}/${nombre}.py &
-    fi
-    ;;
-
-  stop)
-    if [ -f /var/lock/${nombre}.lock ]; then
-      echo "Deteniendo ${nombre}.py"
-      pkill -f ${directorio}/${nombre}.py
-    else
-      echo "${nombre}.py no está en ejecucuón"
-    fi
-    ;;
-
-  restart)
-    /etc/init.d/${nombre} stop && /etc/init.d/${nombre} start
-    ;;
-
-  status)
-    if [ -f /var/lock/${nombre}.lock ]; then
-      echo "${nombre}.py está en ejecución"
-    else
-      echo "${nombre}.py no está en ejecución"
-    fi
-    ;;
-
-  *)
-    echo "Uso: /etc/init.d/${nombre} {start|stop|restart|status}"
-    exit 1
-    ;;
-
-esac
-
-exit 0
+else
+	nombre=reiniciar_router
+	directorio='/opt/RPPGCT'
+	
+	
+	case "$1" in
+	
+		start)
+			if [ -f /var/lock/${nombre}.lock ]; then
+				echo "${nombre}.py ya está en ejecucuón"
+			else
+				echo "Iniciando ${nombre}.py"
+				${directorio}/${nombre}.py &
+			fi
+			;;
+	
+		stop)
+			if [ -f /var/lock/${nombre}.lock ]; then
+				echo "Deteniendo ${nombre}.py"
+				pkill -f ${directorio}/${nombre}.py
+			else
+				echo "${nombre}.py no está en ejecucuón"
+			fi
+			;;
+	
+			restart)
+				/etc/init.d/${nombre} stop && /etc/init.d/${nombre} start
+				;;
+	
+			status)
+				if [ -f /var/lock/${nombre}.lock ]; then
+					echo "${nombre}.py está en ejecución"
+				else
+					echo "${nombre}.py no está en ejecución"
+				fi
+				;;
+	
+			*)
+				echo "Uso: /etc/init.d/${nombre} {start|stop|restart|status}"
+				exit 1
+				;;
+	
+		esac
+	
+	exit 0
